requirements.txt
beautifulsoup4
Flask
Flask-Migrate==1.8.0
Flask-Script==2.0.5
Flask-SQLAlchemy===2.1 
# gunicorn
virtualenv
pycodestyle==2.3.1
psycopg2==2.6.1 
pytest
pytest-cov
pytest-eradicate
pytest-flake8
pytest-flask
pytest-mock
